feat: subdivide api
