+++ package.json
@@ -4,20 +4,22 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "mocha"
+    "test": "mocha",
+    "api": "nodemon ."
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "axios": "^1.6.8",
-    "dotenv": "^16.4.5",
-    "mongoose": "^8.3.2",
-    "passport-local-mongoose": "^8.0.0",
     "bcryptjs": "^2.4.3",
-    "express": "^4.19.2"
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
+    "mongoose": "^8.3.2",
+    "passport-local-mongoose": "^8.0.0"
   },
   "devDependencies": {
     "chai": "^4.3.0",
-    "mocha": "^10.4.0"
+    "mocha": "^10.4.0",
+    "nodemon": "^3.1.0"
   }
 }
